requirements.txt
telethon
aiosqlite
PySocks